+++ src/nvim/if_cscope.c
@@ -1648,7 +1648,7 @@ static void cs_print_tags_priv(char **matches, char **cntxts,
     assert(buf_len >= 0);
 
     // Print the context only if it fits on the same line.
-    if (msg_col + buf_len >= (int)Columns) {
+    if (msg_col + buf_len >= Columns) {
       msg_putchar('\n');
     }
     msg_advance(12);
