+++ src/nvim/os/input.c
@@ -321,10 +321,10 @@ static unsigned int handle_mouse_event(char **ptr, uint8_t *buf,
       // Make sure the mouse position is valid.  Some terminals may
       // return weird values.
       if (col >= Columns) {
-        col = (int)Columns - 1;
+        col = Columns - 1;
       }
       if (row >= Rows) {
-        row = (int)Rows - 1;
+        row = Rows - 1;
       }
       mouse_grid = 0;
       mouse_row = row;
